Consolidate GreyVest ListItem and ListGroupItem
